+++ hassil/parse_expression.py
@@ -85,7 +85,14 @@ def parse_group(
             ParseType.LIST,
             ParseType.RULE,
         ):
-            item = parse_expression(item_chunk, metadata=metadata)
+            # Chunk text ends with explicit whitespace
+            is_end_of_word = (item_chunk.end_index < len(seq_text)) and seq_text[
+                item_chunk.end_index
+            ].isspace()
+
+            item = parse_expression(
+                item_chunk, metadata=metadata, is_end_of_word=is_end_of_word
+            )
 
             if isinstance(grp, (Alternative, Permutation)):
                 # Add to the most recent sequence
@@ -132,7 +139,9 @@ def parse_group(
 
 
 def parse_expression(
-    chunk: ParseChunk, metadata: Optional[ParseMetadata] = None
+    chunk: ParseChunk,
+    metadata: Optional[ParseMetadata] = None,
+    is_end_of_word: bool = True,
 ) -> Expression:
     if chunk.parse_type == ParseType.WORD:
         original_text = _remove_escapes(chunk.text)
@@ -153,7 +162,7 @@ def parse_expression(
     if chunk.parse_type == ParseType.LIST:
         text = _remove_escapes(chunk.text)
         list_name = _remove_delimiters(text, LIST_START, LIST_END)
-        return ListReference(list_name=list_name)
+        return ListReference(list_name=list_name, is_end_of_word=is_end_of_word)
 
     if chunk.parse_type == ParseType.RULE:
         text = _remove_escapes(chunk.text)