+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 2.2.3
+
+- Fix behavior with wildcards inside and outside words
+
+## 2.2.2
+
+- Allow "," as a decimal separator for fractional ranges
+
+## 2.2.1
+
+- Allow list values with "in" but no "out"
+
+## 2.2.0
+
+- Add "fractions" to number ranges with halves and tenths
+- Don't remove punctuation within words in `text_clean` (e.g., "2.5")
+
+## 2.1.1
+
+- Allow number ranges to have the same start/stop (single number)
+
+## 2.1.0
+
+- Upgrade to `unicode-rbnf` 2.2
+- Transition to pyproject.toml
+
 ## 2.0.4
 
 - Trie values are accumulated on `insert` instead of being overwritten